Add CompletionStage support in TarantoolClient
Minimal java version changed from 1.6 to 1.8.

Add support of composable operations as ability to return
TarantoolClientOps with CompletionStage as Result. This feature is taken
from peterservice's fork (see
https://github.com/peterservice-rnd/tarantool-java)

FutureImpl is replaced with CompletableFuture under the hood of
TarantoolClientImpl. Note about behaviour of CompletableFuture
(relatively FutureImpl): a result of a CompletableFuture future will not
change after the first call of the 'complete' method (unlike
FutureImpl.setValue).

test: add ClientComposableAsyncOpsIT and AbstractAsyncClientOperationsIT
as ancestor of it and AsyncClientOperationsIT.

Fixes #102