Support PreparedStatement stream parameters
Add capability to set streams as a parameter of ASCII or UNICODE text as
well as SCALAR binary data. The input streams often are size unaware so
it can be difficult to calculate the data size and, consequently, the
iproto packet size. It leads the driver materializes the streams in
proper data in advance and sends prepared the data as an ordinary text
or raw bytes. Follow API is more important thing now than performance
issues with streams.

Closes: #190