Revert "Change getUnicodeStream encoding from UTF-8 to UTF-16."
This reverts commit d4d62e99.

To be consistent with PreparedStatement.setUnicodeStream which provides
UTF-8 encoding, ResultSet.getUnicodeStream has to also use the same
encoding. Changes in d4d62e99 were made under the influence of JDBC API
for getUnicodeStream that says 'a stream of two-byte 3 characters. The
first byte is the high byte; the second byte is the low byte'.